entropy: Improve types

+++ components/entropy.c
@@ -1,13 +1,14 @@
 /* See LICENSE file for copyright and license details. */
 #if defined(__linux__)
 	#include <stdio.h>
+	#include <stdint.h>
 
 	#include "../util.h"
 
 	const char *
 	entropy(void)
 	{
-		int num;
+		uint64_t num;
 
 		if (pscanf("/proc/sys/kernel/random/entropy_avail",
 		           "%d", &num) != 1) {
